Factuur naar Customerservice

+++ exportpdf.php
@@ -144,7 +144,7 @@ fwrite($myFile, $pdf);
 
 // Verstuur mail
 include("./sendMail.php");
-if (verstuurFactuur($voornaam, $tussenvoegsel, $achternaam, $email, $fileLocation) == TRUE) {
+if (verstuurFactuur($voornaam, $tussenvoegsel, $achternaam, $email, $fileLocation) == TRUE && verstuurFactuurNerdy($voornaam, $tussenvoegsel, $achternaam, $email, $fileLocation) == TRUE) {
     // Verwijder file
     fclose($myFile);
     unlink($fileLocation);
